ParkManager_Fixed/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
channels==4.0.0
channels-redis==4.1.0
pymysql==1.1.0
celery==5.3.4
redis==5.0.1
WeasyPrint==60.1
qrcode==7.4.2
Pillow
django-cors-headers==4.3.1
twilio==8.10.0
reportlab==4.0.7
python-barcode==0.15.1
dj-database-url==2.1.0
python-decouple==3.8
